fixed a bunch warnings and remarks issued by Intel Compiler
git-svn-id: https://lmms.svn.sf.net/svnroot/lmms/trunk/lmms@1722 0778d3d1-df1d-0410-868b-ea421aaaa00d

+++ src/core/journalling_object.cpp
@@ -155,7 +155,7 @@ void journallingObject::changeID( jo_id_t _id )
 					dynamic_cast<automatableModel *>( jo )->
 								displayName();
 			}
-			printf( "JO-ID %d already in use by %s!\n", _id,
+			printf( "JO-ID %d already in use by %s!\n", (int) _id,
 						used_by.toAscii().constData() );
 			return;
 		}
